+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.15.0]
+### Changed
+- More descriptive error messages
+- Disk space error message now tells how much of additional space is needed
+
+
 ## [3.13.0]
 ### Added
 - Speed up unpacking app data with multithreading (#1270)